# src/swagger_server/models/cdr_dataset.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class CdrDataset(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, name=None):
        """
        CdrDataset - a model defined in Swagger

        :param name: The name of this CdrDataset.
        :type name: str
        """
        self.swagger_types = {
            'name': str
        }

        self.attribute_map = {
            'name': 'name'
        }

        self._name = name

    @classmethod
    def from_dict(cls, dikt):
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CdrDataset of this CdrDataset.
        :rtype: CdrDataset
        """
        return deserialize_model(dikt, cls)

    @property
    def name(self):
        """
        Gets the name of this CdrDataset.

        :return: The name of this CdrDataset.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name):
        """
        Sets the name of this CdrDataset.

        :param name: The name of this CdrDataset.
        :type name: str
        """

        self._name = name