# src/swagger_server/models/cdr_individual.py
# coding: utf-8

from __future__ import absolute_import
from swagger_server.models.labels import Labels
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class CdrIndividual(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, name=None, labels=None):
        """
        CdrIndividual - a model defined in Swagger

        :param name: The name of this CdrIndividual.
        :type name: str
        :param labels: The labels of this CdrIndividual.
        :type labels: Labels
        """
        self.swagger_types = {
            'name': str,
            'labels': Labels
        }

        self.attribute_map = {
            'name': 'name',
            'labels': 'labels'
        }

        self._name = name
        self._labels = labels

    @classmethod
    def from_dict(cls, dikt):
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CdrIndividual of this CdrIndividual.
        :rtype: CdrIndividual
        """
        return deserialize_model(dikt, cls)

    @property
    def name(self):
        """
        Gets the name of this CdrIndividual.
        Unique within a dataset.

        :return: The name of this CdrIndividual.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name):
        """
        Sets the name of this CdrIndividual.
        Unique within a dataset.

        :param name: The name of this CdrIndividual.
        :type name: str
        """

        self._name = name

    @property
    def labels(self):
        """
        Gets the labels of this CdrIndividual.

        :return: The labels of this CdrIndividual.
        :rtype: Labels
        """
        return self._labels

    @labels.setter
    def labels(self, labels):
        """
        Sets the labels of this CdrIndividual.

        :param labels: The labels of this CdrIndividual.
        :type labels: Labels
        """

        self._labels = labels